// FRONTEND/chatbot-admin-panel/src/services/api/schemas/roleCreate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { RoleCreateDescription } from './roleCreateDescription';

export interface RoleCreate {
  /**
   * Nombre único del rol (ej. superadmin, context_editor)
   * @minLength 3
   * @maxLength 50
   */
  name: string;
  /** Descripción detallada del rol. */
  description?: RoleCreateDescription;
}